Seperated neuron and weight error models
